requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
pydantic
pydantic-settings
faiss-cpu
sentence-transformers
groq
pytest
httpx
pytest-asyncio
